KnotPoint type

A common way of dealing with trajectories of forced dynamical systems, especially in optimization, is to represent a trajectory with a fixed number of "knot points", typically distributed evenly over time. Each point records the states, controls, time, and time step to the next point. It is often convenient to store all this information together, which is the purpose of the AbstractKnotPoint type. Additionally, it is almost always more efficient to index into a concatenated vector than it is to concatenate two smaller vectors, so the states and controls are stacked together in a single n+m-dimensional vector.

RobotDynamics.jl defines a couple different implementations of the AbstractKnotPoint interface, which can be useful depending on the application.

Types

RobotDynamics.AbstractKnotPoint — Type

AbstractKnotPoint{Nx,Nu,V,T}

Stores the states, controls, time, and time step at a single point along a trajectory of a forced dynamical system with n states and m controls.

Interface

All instances of AbstractKnotPoint should support the following methods:

state_dim(z)::Integer         # state vector dimension
control_dim(z)::Integer       # control vector dimension
getparams(z)                  # arbitrary parameters, usually the time and time step 
getdata(z)::V                 # the underlying vector

Missing docstring.

Missing docstring for GeneralKnotPoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for KnotPoint. Check Documenter's build log for details.

Missing docstring.

Missing docstring for StaticKnotPoint. Check Documenter's build log for details.

Methods

All AbstractKnotPoint types support the following methods:

Missing docstring.

Missing docstring for state. Check Documenter's build log for details.

Missing docstring.

Missing docstring for control. Check Documenter's build log for details.

Missing docstring.

Missing docstring for is_terminal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_z. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_state!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_control!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for set_z!. Check Documenter's build log for details.

Mathematical Operations

All AbstractKnotPoint types support addition between two knot points, addition of a knot point and a vector of length n+m, and multiplication with a scalar, all of which will return a StaticKnotPoint.